test: consolidate end-to-end suites
