+++ .github/workflows/meet-e2e.yml
@@ -7,6 +7,11 @@ on:
       - main
     paths:
       - "suite/meet/**"
+      - "suite/e2e/meet/**"
+      - "suite/e2e/shared/**"
+      - "suite/e2e/package.json"
+      - "suite/e2e/yarn.lock"
+      - "suite/e2e/tsconfig.json"
       - "frontend/src/apps/meet/**"
       - "package.json"
       - "yarn.lock"
@@ -15,6 +20,11 @@ on:
   pull_request:
     paths:
       - "suite/meet/**"
+      - "suite/e2e/meet/**"
+      - "suite/e2e/shared/**"
+      - "suite/e2e/package.json"
+      - "suite/e2e/yarn.lock"
+      - "suite/e2e/tsconfig.json"
       - "frontend/src/apps/meet/**"
       - "package.json"
       - "yarn.lock"
@@ -148,7 +158,7 @@ jobs:
           yarn build
 
       - name: Install E2E Dependencies
-        working-directory: ${{ github.workspace }}/suite/meet/e2e
+        working-directory: ${{ github.workspace }}/suite/e2e
         timeout-minutes: 20
         run: |
           yarn install --frozen-lockfile --non-interactive
@@ -183,8 +193,8 @@ jobs:
           SOCKET_PING_TIMEOUT: 3000
 
       - name: Run E2E Tests
-        working-directory: ${{ github.workspace }}/suite/meet/e2e
-        run: yarn test --shard=${{ matrix.shardIndex }}/${{ matrix.shardTotal }}
+        working-directory: ${{ github.workspace }}/suite/e2e
+        run: yarn test:meet --shard=${{ matrix.shardIndex }}/${{ matrix.shardTotal }}
         env:
           BASE_URL: http://meet.test:8000
           SFU_URL: http://localhost:3000
@@ -198,8 +208,8 @@ jobs:
         with:
           name: meet-playwright-report-shard-${{ matrix.shardIndex }}
           path: |
-            suite/meet/e2e/playwright-report/
-            suite/meet/e2e/test-results/
-            suite/meet/e2e/results.xml
+            suite/e2e/meet/playwright-report/
+            suite/e2e/meet/test-results/
+            suite/e2e/meet/results.xml
           compression-level: 1
           retention-days: 7
