+++ package.json
@@ -8,10 +8,10 @@
   "scripts": {
     "build": "yarn workspace @suite/frontend build",
     "dev": "yarn workspace @suite/frontend dev",
-    "test:e2e": "yarn --cwd suite/meet/e2e test",
-    "test:e2e:drive-writer": "yarn --cwd suite/e2e test",
-    "knip:e2e": "knip --workspace suite/meet/e2e",
-    "knip:e2e:drive-writer": "knip --workspace suite/e2e",
+    "test:e2e": "yarn --cwd suite/e2e test",
+    "test:e2e:productivity": "yarn --cwd suite/e2e test:productivity",
+    "test:e2e:meet": "yarn --cwd suite/e2e test:meet",
+    "knip:e2e": "knip --workspace suite/e2e",
     "knip:frontend": "knip --workspace frontend",
     "knip:sfu-server": "knip --workspace suite/meet/sfu-server",
     "knip:sheets-collab": "knip --workspace suite/sheets/collab-server",
