+++ suite/e2e/meet/specs/restricted-meeting.spec.ts
