--- suite/meet/e2e/.gitignore
@@ -1,6 +1,0 @@
-node_modules/
-playwright-report/
-test-results/
-results.xml
-.state/
-.DS_Store
