--- suite/meet/e2e/package.json
@@ -1,18 +1,0 @@
-{
-	"name": "meet-e2e-tests",
-	"private": true,
-	"version": "0.1.0",
-	"scripts": {
-		"test": "playwright test",
-		"test:headed": "playwright test --headed",
-		"test:debug": "playwright test --debug",
-		"test:ui": "playwright test --ui",
-		"test:report": "playwright show-report",
-		"typecheck": "tsc --noEmit"
-	},
-	"devDependencies": {
-		"@playwright/test": "^1.53.0",
-		"@types/node": "^22.15.3",
-		"typescript": "^5.8.3"
-	}
-}
