--- suite/meet/e2e/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ES2022",
-		"module": "NodeNext",
-		"moduleResolution": "NodeNext",
-		"strict": true,
-		"esModuleInterop": true,
-		"skipLibCheck": true,
-		"resolveJsonModule": true,
-		"types": ["node", "@playwright/test"],
-		"baseUrl": "."
-	},
-	"include": ["./**/*.ts"]
-}
